更新文档和代码：修正README中的链接，调整深度和颜色图像导出函数的默认参数，优化命令行参数解析，确保文件存在性检查，修复训练数据集参数

+++ scannet_wrangling_scripts/SensorData.py
@@ -95,7 +95,7 @@ class SensorData:
       self.num_IMU_frames = struct.unpack('Q', f.read(8))[0]
 
 
-  def export_depth_images(self, output_path, image_size=None, frame_skip=1):
+  def export_depth_images(self, output_path, image_size=(224, 224), frame_skip=1):
     if not os.path.exists(output_path):
       os.makedirs(output_path)
     print('exporting', len(self.frames), frame_skip, ' depth frames to', output_path)
@@ -105,16 +105,25 @@ class SensorData:
 
       if image_size is not None:
         depth = cv2.resize(depth, (image_size[0], image_size[1]), interpolation=cv2.INTER_NEAREST)
-        filepath = os.path.join(output_path, f"frame-{f:06d}.depth.{int(image_size[0])}.png")
+        filepath = os.path.join(output_path, f"frame-{f:06d}.depth.{int(image_size[0])}.png") # 
+        # remove old file if exists
+        # filepath_old = os.path.join(output_path, f"frame-{f:06d}.depth.png")
+        # if os.path.exists(filepath_old):
+        #   try:
+        #       os.remove(filepath_old)
+        #   except FileNotFoundError:
+        #       print(f"File {filepath_old} not found for removal.")
       else:
         filepath = os.path.join(output_path, f"frame-{f:06d}.depth.png")
-
+      
+      if os.path.exists(filepath):
+          return
       with open(filepath, 'wb') as f: # write 16-bit
         writer = png.Writer(width=depth.shape[1], height=depth.shape[0], bitdepth=16)
         depth = depth.reshape(-1, depth.shape[1]).tolist()
         writer.write(f, depth)
 
-  def export_color_images(self, output_path, image_size=None, frame_skip=1):
+  def export_color_images(self, output_path, image_size=(224, 224), frame_skip=1):
     if not os.path.exists(output_path):
       os.makedirs(output_path)
     print('exporting', len(self.frames), frame_skip, 'color frames to', output_path)
@@ -122,19 +131,31 @@ class SensorData:
       color = self.frames[f].decompress_color(self.color_compression_type)
 
       if image_size is not None:
-        resized = Image.fromarray(color).resize((image_size[0], image_size[1]), resample=Image.BILINEAR)
         filepath = os.path.join(output_path, f"frame-{f:06d}.color.{int(image_size[0])}.png")
+        if os.path.exists(filepath):
+          return
+        resized = Image.fromarray(color).resize((image_size[0], image_size[1]), resample=Image.BILINEAR)
         resized.save(filepath)
+        # old_fp = os.path.join(output_path, f"frame-{f:06d}.color.jpg")
+        # if os.path.exists(old_fp):
+        #   try:
+        #       os.remove(old_fp)
+        #   except FileNotFoundError:
+        #       print(f"File {old_fp} not found for removal.")
       else:
         filepath = os.path.join(output_path, f"frame-{f:06d}.color")
+        if os.path.exists(filepath) or  os.path.exists(filepath+".jpg"):
+          return 
         self.frames[f].dump_color_to_file(self.color_compression_type, filepath)
 
 
   def save_mat_to_file(self, matrix, filename):
+    # if file exists, not overwrite
+    if os.path.exists(filename):
+      return
     with open(filename, 'w') as f:
       for line in matrix:
         np.savetxt(f, line[np.newaxis], fmt='%f')
-
 
   def export_poses(self, output_path, frame_skip=1):
     if not os.path.exists(output_path):
